openssl: implement cgoless wrappers for C functions (#288)
* openssl: implement cgoless wrappers for C functions

* fix aes test

* remove darwin pins

* unskip md4 hash test

* enable more tests

* re enable symcrypt tests

* add boiletplate linux syscall file

* auto generate variadic functions

* add checkheader if statement back

* fix windows

* implement windows support

* fix iSAvailable functions

* fix renamed function in nocgo

* support dynload mode

* panic if non-optional functions can't be loaded

* deduplicate code

* fix unix

* fix windows

* fix windows

* fix windows

* deduplicate variadic handling

* simplify code a bit

* fix nocgo error handling

* fix typedefs

* fix dlsym

* fix pointer tests

* several improvements and fixes

* fix cTypeSize

* fix type map

* fix type map

* fix data types

* use dlclose on nocgo

* fix newMkcgoErr

* unhardcode dlopen and friends

* fix RTLD_LOCAL

* use keepalive

* support cgoless linux and darwin

* add cgoless test to CI

* gate syscallN assembly

* deduplicate some code

* revert msgo bump

* mkcgo cleanup

* fix static trmapolines

* fix static trmapolines

* remove point_conversion_form_t special case

* remove point_conversion_form_t special case

* remove stale comment

* don't hardcode zdl.go

* don't hardcode zdl.go

* fix needsAssembly

* cleanup

* deduplicate fakecgo

* support 32-bit arches

* fix fakecgo

* gate ocgo changes behind a build flag

* readd cgoless without build tag test

* add eof

* revert changes in the root package

* add missing build tags

* add cgo suffix to cgo files

* put common nocgo and cgo code in same file

* add missing build tags

* don't generate 1.24 file if not needed

* generalize checkheader

* add copyright flag to mkcgo

* Update internal/ossl/errors_nocgo.go


---------

+++ internal/ossl/errors_cgo.go
@@ -81,7 +81,7 @@ func newMkcgoErr(msg string, errst C.mkcgo_err_state) error {
 	b.WriteString(msg)
 	b.WriteString("\nopenssl error(s):")
 	for i := range C.ERR_NUM_MAX {
-		e := uint32(oerrst.code[i])
+		e := uint64(oerrst.code[i])
 		if e == 0 {
 			break
 		}
